+++ dessia_common/core.py
@@ -13,7 +13,7 @@ import collections
 from copy import deepcopy
 import inspect
 import json
-from typing import List, Tuple, Union, Any, \
+from typing import List, Dict, Type, Tuple, Union, Any, \
     get_type_hints, get_origin, get_args
 try:
     from typing import TypedDict  # >=3.8
@@ -185,11 +185,11 @@ class DessiaObject:
         if full_classname(self) != full_classname(other_object):
             return False
 
-        eq_dict = self.__getstate__()
+        eq_dict = self._serializable_dict()
         if 'name' in eq_dict:
             del eq_dict['name']
             
-        other_eq_dict = other_object.__getstate__()
+        other_eq_dict = other_object._serializable_dict()
 
         for key, value in eq_dict.items():
             other_value = other_eq_dict[key]
@@ -202,7 +202,7 @@ class DessiaObject:
         forbidden_keys = (self._non_data_eq_attributes
                           + self._non_data_hash_attributes
                           + ['package_version', 'name'])
-        for key, value in self.__getstate__().items():
+        for key, value in self._serializable_dict().items():
             if key not in forbidden_keys:
                 if isinstance(value, list):
                     hash_ += list_hash(value)
@@ -225,9 +225,9 @@ class DessiaObject:
         # eq_dict = {k: v for k, v in self.to_dict().items()
         #            if (k not in ['package_version', 'name'])\
         #                and (k not in self._non_data_eq_attributes)}
-        eq_dict = self.__getstate__()
+        eq_dict = self._serializable_dict()
         # other_eq_dict = other_object.to_dict()
-        other_eq_dict = other_object.__getstate__()
+        other_eq_dict = other_object._serializable_dict()
 
         for key, value in eq_dict.items():
             if key not in other_eq_dict:
@@ -260,17 +260,19 @@ class DessiaObject:
             dict_['package_version'] = package_version
         return dict_
 
-    def __getstate__(self):
+    def _serializable_dict(self):
+
         dict_ = {k: v for k, v in self.__dict__.items()
                  if k not in self._non_serializable_attributes
                  and not k.startswith('_')}
         return dict_
 
+
     def to_dict(self) -> JsonSerializable:
         """
         Generic to_dict method
         """
-        dict_ = self.__getstate__()
+        dict_ = self._serializable_dict()
         if hasattr(self, 'Dict'):
             # !!! This prevent us to call DessiaObject.to_dict()
             # from an inheriting object which implement a Dict method,
@@ -339,6 +341,16 @@ class DessiaObject:
 
         unordered_count = 0
 
+        # Parse docstring
+        try:
+            parsed_docstring = parse_docstring(cls)
+        except Exception:
+            parsed_docstring = {
+                'description': 'Docstring parsing failed',
+                'attributes': {}
+            }
+        parsed_attributes = parsed_docstring['attributes']
+
         # Initialize jsonschema
         _jsonschema = deepcopy(JSONSCHEMA_HEADER)
 
@@ -346,6 +358,7 @@ class DessiaObject:
                                                                   merge=False)
         _jsonschema['required'] = required_arguments
         _jsonschema['standalone_in_db'] = cls._standalone_in_db
+        _jsonschema['description'] = parsed_docstring['description']
 
         # Set jsonschema
         for annotation in annotations.items():
@@ -364,11 +377,15 @@ class DessiaObject:
                 editable = name not in cls._non_editable_attributes
                 annotation_type = type_from_annotation(annotation[1], cls)
                 annotation = (annotation[0], annotation_type)
-                jss_elt = jsonschema_from_annotation(annotation=annotation,
-                                                     jsonschema_element={},
-                                                     order=order,
-                                                     editable=editable,
-                                                     title=title)
+                jss_elt = jsonschema_from_annotation(
+                    annotation=annotation, jsonschema_element={},
+                    order=order, editable=editable, title=title
+                )
+                if name in parsed_attributes:
+                    description = parsed_attributes[name]['desc']
+                    typing_ = parsed_attributes[name]['annotation']
+                    jss_elt[name].update({'description': description,
+                                          'python_typing': typing_})
                 _jsonschema['properties'].update(jss_elt)
                 if name in default_arguments.keys():
                     default = set_default_value(_jsonschema['properties'],
@@ -1726,3 +1743,41 @@ def default_dict(jsonschema):
     else:
         return None
     return dict_
+
+
+class ParsedAttribute(TypedDict):
+    desc: str
+    type_: str
+    annotation: str
+
+
+class ParsedDocstring(TypedDict):
+    description: str
+    attributes: Dict[str, ParsedAttribute]
+
+
+def parse_docstring(cls: Type) -> ParsedDocstring:
+    """
+    Parse docstring of given class. Refer to docs to see how docstrings
+    should be built.
+    """
+    annotations = get_type_hints(cls.__init__)
+    docstring = cls.__doc__
+    if docstring:
+        splitted_docstring = docstring.split(':param ')
+        parsed_docstring = {"description": splitted_docstring[0].strip()}
+        params = splitted_docstring[1:]
+        args = {}
+        for param in params:
+            splitted_param = param.split(':type ')
+            arg = splitted_param[0]
+            typestr = splitted_param[1]
+            argname, argdesc = arg.split(":", maxsplit=1)
+            argtype = typestr.split(argname+":")[-1]
+            annotation = annotations[argname]
+            args[argname] = {'desc': argdesc.strip(), 'type_': argtype.strip(),
+                             'annotation': str(annotation)}
+            # TODO Should be serialize typing ?
+        parsed_docstring.update({'attributes': args})
+        return parsed_docstring
+    return {'description': "", 'attributes': {}}