+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased[=>0.4.3]
+## Unreleased => [0.4.4]
+### Added
+- Docstring parsing & failure prevention
+- Description of class and attributes in jsonschema
+
+## [0.4.3]
 ### Added
 - Datatype from jsonschema method
 - Method flag in jsonschema